Refactored Sspi.Win32Exception and LibgssapiException to derive from GssapiException to simplify exception handling.
